requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-rest-passwordreset==1.4.0
djangorestframework==3.15.1
PyJWT==2.8.0
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
